+++ src/Workflowish/Item/ChildItems.tsx
@@ -3,10 +3,13 @@ import { ItemTreeNode, TransformedDataAndSetter } from "~Workflowish/mvc/model";
 import Item, { FocusActions, ItemRef, ItemStyleParams } from ".";
 import { ControllerActions, makeItemActions } from "../mvc/controller"
 import { FocusedActionReceiver } from "~Workflowish/mvc/focusedActionReceiver";
+import { TreePath, DFSFocusManager } from "~Workflowish/mvc/DFSFocus";
 
 export const ChildItems = (props: {
     shouldUncollapse: boolean,
     item: ItemTreeNode,
+    treePath: TreePath,
+    focusManager: DFSFocusManager,
     styleParams: ItemStyleParams,
     itemsRefArray: React.MutableRefObject<(FocusActions | null)[]>,
     setThisAsFocused: (focusedActionReceiver: FocusedActionReceiver, focusItemKey: string) => void,
@@ -35,8 +38,8 @@ export const ChildItems = (props: {
                 borderLeft: "1px solid white",
                 marginLeft: "0.5em"
             }}>
-                {childrenToRender.map((item, ii) => (<Item
-                    key={ii}
+                {childrenToRender.map((item, childIdx) => (<Item
+                    key={childIdx}
                     item={item}
                     styleParams={{
                         showId: props.styleParams.showId,
@@ -45,9 +48,12 @@ export const ChildItems = (props: {
                     setThisAsFocused={props.setThisAsFocused}
                     actions={makeItemActions({
                         thisItem: item,
-                        focusItem: props.actions.focusItem,
+                        treePath: props.focusManager.childPath(props.treePath, childIdx),
+                        focusManager: props.focusManager,
                         model: props.model
                     })}
+                    treePath={props.focusManager.childPath(props.treePath, childIdx)}
+                    focusManager={props.focusManager}
                     model={props.model}
                     pushRef={props.pushRef}
                 ></Item>))}