+++ src/Workflowish/index.tsx
@@ -9,6 +9,7 @@ import { FloatyButtons } from "./Subcomponents/FloatyButtons";
 import OmnibarWrapper from "./Subcomponents/OmnibarWrapper";
 import ContextMenu from "./Subcomponents/ContextMenu";
 import { ModelContext, RenderTimeContext } from "./mvc/context";
+import { DFSFocusManager } from "./mvc/DFSFocus";
 
 export default (props: {
     data: BaseStoreDataType,
@@ -69,12 +70,12 @@ const ItemsList = (
     }
 ) => {
     const itemTree = React.useContext<ItemTreeNode>(ModelContext);
-
+    const focusManager = new DFSFocusManager();
     const setFocusedItem = (focusedActionReceiver: FocusedActionReceiver, focusItemKey: string) => {
         props.setFocusedActionReceiver(focusedActionReceiver);
         props.setLastFocusedItem(focusItemKey);
     }
-
+    const rootPath = [0];
     return <RenderTimeContext.Provider value={{
         currentFocusedItem: props.lastFocusedItem
     }}>{itemTree.children.map((item, ii) => {
@@ -88,13 +89,14 @@ const ItemsList = (
             pushRef={(id: string, ref: ItemRef) => props.itemRefsDictionary[id] = ref}
             setThisAsFocused={setFocusedItem}
             actions={makeItemActions({
-                focusItem: (id: string) => {
-                    props.itemRefsDictionary[id].focusThis();
-                },
+                focusManager: focusManager,
+                treePath: focusManager.childPath(rootPath, ii),
                 disableDelete: () => (itemTree.children.length == 1),
                 thisItem: item,
                 model: props.transformedDataAndSetter,
             })}
+            focusManager={focusManager}
+            treePath={focusManager.childPath(rootPath, ii)}
             model={props.transformedDataAndSetter}
         ></Item >)
     })}</RenderTimeContext.Provider>