+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # Latest 
 - Add CLI cleanup tools
 - first round of top-down-update code removal
+- fix focus traversal and child creation
 
 # Version 3.6.0: Add omnibar 
 - Refactor to separate omnibar and search